todo list, and choice of format;
I'm foregoing the standard readme template from lein.  I'm really only
doing this for myself, not for the sake of collaboration.  So I'm not
really worried about having a recognized format.  When there are docs,
they will be more stream of consciousness.  Having said that, people are
of-course still welcome to fork and tinker.

I also added a ToDo list, as a motivational tactic, and to help chip
away at it.