new tests, and supporting scaffolding;
I used the utility functions to generate the test data, and I built new
tests which use that test data.  But those tests are failing, but I'm
not sure if it's because the tests are wrong, or they've found a flaw in
my logic (I feel like it may be the latter).